// src/interfaces/ContatoInterface.tsx
export interface IContato {
    id: number;
    nome: string;
    cpf_cnpj: string;
    endereco: string;
    numero: string;
    cep: string;
    complemento: string;    
    bairro: string;
    cidade: string;
    estado: string;
    email: string;
    telefone: string;
    celular: string;
    ativo: boolean;
}